Create: Details page, showcase, needs propose functionality & styling Fix to do: logo in header should not act as h1.
commented out middleware as it was conflicting with details right now.
did not want to change it for now before creating login.

minor alt fixes on images in header, footer

updated listingsprovider to be more advanced, also for fixing reloads on pagination, so now it only stores it in url for share/bookmarks, instead of driving the state, the state is handled only on client.